requirements.txt
pandas
numpy
openpyxl
sqlalchemy
pymysql
python-dotenv
scikit-learn
prophet
matplotlib
seaborn
fastapi
uvicorn[standard]
